+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.2.0 (2020-??-??)
 
 * Added support for essay questions.
-
+* Added support for question groups.
 
 ## v0.1.0 (2020-04-01)